// frontend/src/api/models/handlerSetTenantPlanRequestPlan.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerSetTenantPlanRequestPlan =
  (typeof HandlerSetTenantPlanRequestPlan)[keyof typeof HandlerSetTenantPlanRequestPlan]

export const HandlerSetTenantPlanRequestPlan = {
  free: 'free',
  basic: 'basic',
  pro: 'pro',
  enterprise: 'enterprise',
} as const